Interpretation refactor: re-organize files/classes/functions
