+++ core/src/test/scala/modelfind/NextInterpretationTest.scala
@@ -1,0 +1,5 @@
+package modelfind
+
+class NextInterpretationTest {
+  
+}
